+++ lib/dicey/cli/formatters/gnuplot_formatter.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "key_value_formatter"
+require_relative "base_list_formatter"
 
 module Dicey
   module CLI
@@ -8,7 +8,9 @@ module Dicey
       # Formats a hash as a text file suitable for consumption by Gnuplot.
       #
       # Will transform Rational probabilities to Floats.
-      class GnuplotFormatter < KeyValueFormatter
+      # Non-numeric dice inherently won't work in gnuplot,
+      # even though the formatter will process them.
+      class GnuplotFormatter < BaseListFormatter
         SEPARATOR = " "
 
         private