+++ lib/dicey/cli/formatters/json_formatter.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-require_relative "hash_formatter"
+require_relative "base_map_formatter"
 
 module Dicey
   module CLI
     module Formatters
       # Formats a hash as a JSON document under +results+ key, with optional +description+ key.
-      class JSONFormatter < HashFormatter
+      class JSONFormatter < BaseMapFormatter
         METHOD = :to_json
       end
     end