+++ lib/dicey/cli/formatters/list_formatter.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-require_relative "key_value_formatter"
+require_relative "base_list_formatter"
 
 module Dicey
   module CLI
     module Formatters
       # Formats a hash as list of key => value pairs, similar to a Ruby Hash.
-      class ListFormatter < KeyValueFormatter
+      class ListFormatter < BaseListFormatter
         SEPARATOR = " => "
       end
     end